fixme: move snapfrompid into osutils
