+++ usersession/userd/export_test.go
@@ -23,10 +23,6 @@ import (
 	"github.com/godbus/dbus"
 )
 
-var (
-	SnapFromPid = snapFromPid
-)
-
 func MockSnapFromSender(f func(*dbus.Conn, dbus.Sender) (string, error)) func() {
 	origSnapFromSender := snapFromSender
 	snapFromSender = f
